feat: add P command to play music
